Support var_root in multi events - 2
var_root is used to optionally lift up the keys in a nested events
dictionary.
The var_root has been modified to take in a sigle nested key in the form
of a.b.c or or a list of keys of the form
- a.b.c
- x.y.x

When we have matching on multiple events from different sources the
event payload could have different keys. The var_root can store the
nested keyname for different payloads and pick them from different
events.

If the var_root has the following keys
   ['webhook.payload', 'kafka.message']

it can modify the events from

{'events': {'webhook': {'webhook': {'payload': {'url': 'http://www.example.com', 'action': 'merge'}}},
            'kafka': {'kafka': {'message': {'topic': 'testing', 'channel': 'red'}}}}

to

{'events': {'webhook': {'url': 'http://www.example.com', 'action': 'merge'},
            'kafka': {'topic': 'testing', 'channel': 'red'}}