[AAP-4440] Support storing facts per host
A ruleset can have an optional gather_facts flag which enables
collection of ansible facts per host in the inventory list.
These ansible facts are stored in the rules engine per host.
When a rule matches the corresponding host fact and the event
are returned from the rules engine.

You can refer to the facts per host using the fact prefix.
e.g.

fact.meta.hosts == 'localhost'

When the fact is returned from the rules engine it allows us to limit
the hosts that the playbook would run on by using the meta.hosts key.

https://issues.redhat.com/browse/AAP-4440