Fixed the example of how to use fact and events
The fact and events can't be compared in a simple event.
The fact has to be first assigned to a temporary events field.
Then the saved events field can be compared with the event.

This change was introduced in
https://github.com/ansible/ansible-rulebook/commit/eb4110b1eaadd64e3972661a4d0ca6921143208b

But the document was not fixed.
This PR fixes the documentation