Support for standalone booleans in conditions
This helps when we want to print events without knowing
the contents of the event.
The literal boolean true allows the condition to pass
and the action print_event can be used to print the event

This would usually be used for bulding the rulebook, you would
not use it in production.

```
    - name: print event always
      condition: true
      action:
        print_event:
          pretty: true
```