AAP-6908 update and refactor docs (#277)
How to test

```
cd docs
make clean && make html
cd _build/html
python -m http.server 3000
```
Open the browser: http://localhost:3000

Closes AAP-6908

Includes:
- A new organization that (I hope) is clearer and intuitive. Moves
content from the root project to /docs, leaving the readme.rst as an
entrypoint for everything else. Moves several pieces of content to their
respective sections.
- Links to the ansible code of conduct
- Use code blocks for all code snippets
- Multiple fixes and clarifications
- Some additions
- Removes history since it's not maintained (a proper changelog can be
implemented in the future)


Of course, it can be better. IMO we must add more and useful examples.
There are pages work in progress and probably some definitions can be
clearer. This is something that can be done in the next iterations.

Co-authored-by: Tom Tuffin <71447672+ttuffin@users.noreply.github.com>