Added a echo action
The debug/print_event have lot of information and don't allow for
terse messages. When doing a demo instead of seeing a lot of attributes
flow by we can echo terse messages along with a time stamp.
The timestamp allows for timer based rules to see if they popped at
the expected times.