[AAP-8555] Error parsing not condition
When using multiple logical statments in a condition with or/and
we need ( and ) to resolve ambiguities. Otherwise we should be able
to parse simple conditions with a not without braces. The not has
a priority over the and and or.

e.g. all of these are valid

```
   condition: not event.i == 5
```

```
   condition: not event.msg == "Barney"
```

```
   condition: not event.j >= 10
```

```
   condition: not event.my_boolean
```

But if you use and/or you need surrounding round braces
```
   condition: not (event.i > 50 or event.i < 10)
```