Vars related cleanup
Since we have the vars namespace for accessing variables passed in
from the vars-file or env_vars we don't have to send the vars to the
rule engine. The vars are resolved before we send the ruleset to Drools.
Added additional notes in documentation. Cleaned up tests that were
using facts in conditions to use vars instead