Allow for empty payloads
If the payload is empty then skip over it, in some special cases
the ruleset might not need any direct events but would be using
set_facts to generate the data.
This came about in PR https://github.com/ansible/ansible-rulebook/pull/349
Where there is no payload. This leads to an error in the source
```
2023-02-14 16:38:53,582 - ansible_rulebook.engine - ERROR - Source error
Traceback (most recent call last):
  File "/Users/madhukanoor/devsrc/ansible-rulebook/ansible_rulebook/engine.py", line 154, in start_source
    await entrypoint(fqueue, args)
  File "/Users/madhukanoor/devsrc/ansible-rulebook/tests/e2e/../sources/generic.py", line 97, in main
    data.update(event)
```