Fix intermittently failing run_playbook test
With the introduction of ordered shutdown in PR 358,
an intermittent race condition arose with the run_playbook
test due to Ruleset 2 bailing before it received the set_fact.
This change moves the set_fact to the secondary ruleset so
that it occurs early on in the run, in order to avoid the
race condition.