Fix intermittently failing run_playbook test (#374)
Depends on #358.

With the introduction of ordered shutdown in PR 358, an intermittent
race condition arose with the run_playbook test due to Ruleset 2 bailing
before it received the set_fact from ruleset 1. PR 358 adds ordered
shutdown, but if there are no pending items in the action queue then the
ruleset will end immediately. In this test, once the time specified
under `shutdown_after` in ruleset 1 has expired then a shutdown command
will be issued to all rulesets, and since ruleset 2 does not have any
pending items it will end immediately, sometimes before the `set_fact`
from ruleset 1 has had a chance to run. Increasing `shutdown_after` is
an option, but this increases the test time unnecessarily.

This change moves the condition that performs the set_fact so that it
occurs at the beginning of the run, so we can be sure that ruleset 2
will not end before the set_fact and therefore avoid the race condition.
We may run into a race condition with other arch's, such as ppc, in
which case we would need to increase `SOURCE_SHUTDOWN_AFTER` for that
particular arch.