Support inventory in ini/yaml formats
ansible-rulebook doesn't do anything with the inventory
data. It just reads it and writes into the inventory file
for ansible_runner. So instead of treating every inventory
as a yaml file and load it as a dictionary, this PR just
loads the inventory data as a blob irrespective of the
format and sends it to ansible-runner.

Fixes #241