Fixed worker mode and activation id in parse args
In an earlier fix we had made -r a required parameter which
breaks when used in worker mode. In worker mode there is no rulebook
the rulebook comes in as a payload on the websocket channel.
The --id parameter is a string and not an integer since it could
be the activation guid from the database

This undoes the code changes in
PR #431