Fix intermittent test failures
In the once_after testcase I had the event burst every 15 seconds
and the time window was also 15 seconds so once in a while the
new event burst would get caught in the earlier time window and
get suppressed. Now I have changed the loop delay to 18 seconds
so the event burst happens after 18 seconds.