[AAP-11337] Send heartbeat on web socket channel
If ansible-rulebook is invoked with --heartbeat nn, where nn is
the seconds to wait between sending heartbeat, it will send session
stats about all running ruleset sessions. This allows the server
to monitor if the ansible-rulebook is responsive.

The payload coming back from the server
```
   type: SessionStats
   stats:
       'eventsMatched': 1,
       'eventsProcessed': 5,
       'eventsSuppressed': 4,
       'numberOfDisabledRules': 0,
       'numberOfRules': 1,
       'permanentStorageSize': 0,
       'ruleSetName': '02 Debug',
       'rulesTriggered': 1,
       'sessionId': 1,
       'start': '2023-04-25T13:54:09.448258Z'
       'end': ....
   activation_id:
   reported_at:
```
The end is sent only when the ruleset session has ended