[AAP-12909] Validate Controller URL and Token at Startup
If the caller has provided a controller URL and token we validate
that it is correct at startup so we dont have to fail
when the events arrive if the url or token is invalid.

At startup we get the config information from the Controller using
/api/v2/config and log the version of the controller we are using.

https://issues.redhat.com/browse/AAP-12909

The issue mentions failing on audit_rule, with this PR we fail as
soon as the process starts. The Activation is marked as failed and
the history has the details for bad url or bad token. This doesn't
have to be replicated for every rule and event. If the URL and token
is bad it will stay bad for the entire run of the activation since
the URL and token cannot be changed when the activation is running.