test: tool to convert rulebook yml to ast format (#545)
ansible-rulebook internally uses code to convert a rulebook from yml
format to a Abstract Syntax Tree (ast) format. This AST is then passed
into the rule engine. The parsing is done using pyparsing in
ansible-rulebook. When writing tests for drools_jpy we need the AST
format, since the code is hidden inside ansible-rulebook its not easy to
access it. This tool exposes that ast conversion feature so we can
convert one rulebook at a time to AST format for testing. If the
rulebook has references to vars and env vars they can be passed into the
converter since it would substitute the variables in the AST.

usage:

python3 tools/convert_to_ast.py [-h] -r RULEBOOK [-e VARS] [-o OUTPUT]
[-E ENV_VARS]