fix: wait for the feedback task to end (#562)
[AAP-13874](https://issues.redhat.com/browse/AAP-13874)

ansible-rulebook uses a websocket channel to send feedback events for
reporting purposes. The feedback task runs asynchronously and might get
shutdown/cancelled before it can post all of the events. This PR ensures
that the feedback task gets sufficient time to post all of the events
back to the server