Handle source tasks and feedback tasks separately
tasks was a variable that contained the return from spawn_sources
as well as the websocket feedback task. These are different things,
and should not be passed into run_rulesets, since the feedback task
is not an event source.

tasks is now used as a list of all tasks, only for the asyncio.gather call