feat: reduce image size further
Reduces the size of the resulting image to around 400 MB.
Uses multi-stage build process. Python dependencies are packaged into one venv. Java JRE is created custom, only keeping the modules Drools needs.
Both the venv and the JRE are copied to the second stage of the build based on stream9-minimal image to reduce the size even further.