feat: option to skip including events into extra_vars
When we unconditionally send event data to the controller, if there
is a survey spec in the workflow template the launch of the workflow
template fails. With this option in a rulebook the user can prevent
event data from being sent by setting
   - include_events: false

The default value is true so existing rulebooks will keep sending
events to controller.