fix: [AAP-45945] Retrieve artifacts from set_fact/set_stats event data (#795)
https://issues.redhat.com/browse/AAP-45945

It was recommended by the core team that we should be using
ansible.builtin.set_stats in the playbook if we have to pass back data
from a playbook. This PR adds support for ansible.builtin.set_stats
This PR also fixes an issue in the processing of
ansible.builtin.set_fact so that we can remain backward compatible. It
is highly recommended that users should be using set_stats instead of
set_fact.

This would help us when we migrate to ansible-core 2.19.