chore: update stable-1.0 CI to run py312 (#855)
Update CI for stable/1.0 branch to run against py312. Also include
changes to dependencies for tests to be able to run with py312.

An extra change to engine.py was needed to fix a failing test, which
seems to have been brought to the surface in py312 due to improvements
in asyncio.

https://issues.redhat.com/browse/AAP-54796